:wrench: 폴더링
변경사항:
- AI 관련 폴더 구조 변경
- AI 관련 스크립트 폴더 이동

코멘트:
- AI 관련 기능은 모두 Dev 폴더에서 작성
