+++ Assets/Dev/AI/Data.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 35e50fa0fc0ddaf48845799ad972290e
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
