+++ Assets/Dev/AI/IndieLINY.AI.asmdef.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 4b903d05d0816764a9808e9880e9886f
+AssemblyDefinitionImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
