+++ Assets/Dev/AI/Script.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: d17a0d4c12901b34aafcc656ee800c9e
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
