+++ Assets/Dev/AI/Script/Action.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: eca0d08a22d1f824cb34b8b0f60e0de3
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
