+++ Assets/Dev/AI/Script/Debug.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: f4cadf7883edd9e409818f3711e59e8c
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
