+++ Assets/Dev/AI/Script/Utils/WaitForSec.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 73443987ecb16d74fbf53f00d6ea8d25
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
