--- Assets/Dev/Script/AI/Data.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 35e50fa0fc0ddaf48845799ad972290e
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
