--- Assets/Dev/Script/AI/Translation.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: bd1f12faea6b51e4892e9cf53e4a8fe9
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
