--- Assets/Dev/Script/AI/Util.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: c2ea5246f19581c4db0fb4c9aef74011
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
