+++ Assets/Libaray/Event/IndieLINY.Event.asmdef
@@ -1,0 +1,16 @@
+{
+    "name": "IndieLINY.Event",
+    "rootNamespace": "",
+    "references": [
+        "GUID:6dfffdeca0f41044ab1e81833b5fc950"
+    ],
+    "includePlatforms": [],
+    "excludePlatforms": [],
+    "allowUnsafeCode": false,
+    "overrideReferences": false,
+    "precompiledReferences": [],
+    "autoReferenced": true,
+    "defineConstraints": [],
+    "versionDefines": [],
+    "noEngineReferences": false
+}
