+++ Assets/Libaray/Event/IndieLINY.Event.asmdef.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: a078dd6e911d68a498fa31c509ed06b0
+AssemblyDefinitionImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
