--- Assets/Libaray/FSM.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: d368cd8fb6fe84a4db7afcbfb5d82608
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
