--- Assets/Libaray/FSM/Data.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 25ca56990d544fb4681f740bb0908af2
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
