--- Assets/Libaray/FSM/Graph.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 6359773e2c8341c4fb7c6db0c239b04c
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
