--- Assets/Libaray/FSM/IndieLINY.AI.asmdef
@@ -1,20 +1,0 @@
-{
-    "name": "IndieLINY.AI",
-    "rootNamespace": "",
-    "references": [
-        "GUID:21b0c8d1703a94250bfac916590cea4f",
-        "GUID:ea715009a4efd4c6cbc85be3ae097dd3",
-        "GUID:315d634a9ac6d460a9515f5a56be6311",
-        "GUID:8c4dd21966739024fbd72155091d199e",
-        "GUID:299beb10465c46f41abeaf10478b4727"
-    ],
-    "includePlatforms": [],
-    "excludePlatforms": [],
-    "allowUnsafeCode": false,
-    "overrideReferences": false,
-    "precompiledReferences": [],
-    "autoReferenced": true,
-    "defineConstraints": [],
-    "versionDefines": [],
-    "noEngineReferences": false
-}
