--- Assets/Libaray/FSM/IndieLINY.AI.asmdef.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: d07ce1f741f24a942afdd981e8f6b806
-AssemblyDefinitionImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
