--- Assets/Libaray/FSM/Node.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: b93fbf55ef8e1d3408bbe309ebd3eaf6
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
