--- Assets/Libaray/FSM/Node/Base.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 89819dcc5a50cc743bf7fcb198c1816c
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
