--- Assets/Libaray/FSM/Util.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: c9993d1ffe1e831488838edf220d39f1
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
